requirements.txt
pypdf
sentence-transformers
faiss-cpu # or faiss-gpu if GPU is available
groq
gradio
pandas
numpy
langchain
python-dotenv